+++ blog/src/main/java/com/example/blog/api/dto/MemberDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 //@RequestBody로 외부에서 데이터를 받는 경우엔 기본생성자 + set메소드를 통해서만 값이 할당
-public class MemberDTO{
+public class MemberDto{
     private String userId;
     private String password;
     private String name;
@@ -29,7 +29,7 @@ public class MemberDTO{
     }
 
     @Builder
-    MemberDTO(String userId, String password , String name, String email, String address){
+    MemberDto(String userId, String password , String name, String email, String address){
         this.userId = userId;
         this.password = password;
         this.name = name;
