+++ blog/src/main/java/com/example/blog/api/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.example.blog.api.domain;
+package com.example.blog.api.entity;
 
 import java.time.LocalDateTime;
