+++ blog/src/main/java/com/example/blog/api/entity/Board.java
@@ -1,10 +1,7 @@
-package com.example.blog.api.domain;
+package com.example.blog.api.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.Table;
 
 import lombok.AccessLevel;
