+++ blog/src/main/java/com/example/blog/api/repository/BoardRepository.java
@@ -2,7 +2,7 @@ package com.example.blog.api.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.blog.api.domain.Board;
+import com.example.blog.api.entity.Board;
 
 public interface BoardRepository extends JpaRepository<Board, Long>{
 }
