+++ blog/src/main/java/com/example/blog/api/servcie/MemberService.java
@@ -2,13 +2,13 @@ package com.example.blog.api.servcie;
 
 import java.util.List;
 
-import com.example.blog.api.dto.MemberDTO;
+import com.example.blog.api.dto.MemberDto;
 
 public interface MemberService {
     
-    void insert(MemberDTO dto);
-    MemberDTO searchById(String id);
-    List<MemberDTO> getAll();
-    void update(MemberDTO dto);
+    void insert(MemberDto dto);
+    MemberDto searchById(String id);
+    List<MemberDto> getAll();
+    void update(MemberDto dto);
     void delete(String id);
 }
