+++ blog/src/main/java/com/example/blog/api/servcie/MemberServiceImpl.java
@@ -2,10 +2,11 @@ package com.example.blog.api.servcie;
 
 import java.util.List;
 
-import com.example.blog.api.dto.MemberDTO;
+import com.example.blog.api.dto.MemberDto;
 import com.example.blog.api.repository.MemberRepository;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import lombok.AllArgsConstructor;
 
@@ -15,26 +16,31 @@ public class MemberServiceImpl implements MemberService {
 
     private MemberRepository memberRepository;
 
+    @Transactional
     @Override
-    public void insert(MemberDTO dto) {
+    public void insert(MemberDto dto) {
         memberRepository.save(dto.toEntity());
     }
 
+    @Transactional
     @Override
-    public MemberDTO searchById(String id) {
+    public MemberDto searchById(String id) {
         return null;
     }
 
+    @Transactional
     @Override
-    public List<MemberDTO> getAll() {
+    public List<MemberDto> getAll() {
         return null;
     }
 
+    @Transactional
     @Override
-    public void update(MemberDTO dto) {
+    public void update(MemberDto dto) {
 
     }
 
+    @Transactional
     @Override
     public void delete(String id) {
